Outsource geocoding data in a new file
