+++ tests/geocoding_tests.py
@@ -21,13 +21,13 @@ from sqlalchemy.engine import reflection
 from superset import db
 from superset.connectors.sqla.models import SqlaTable
 from superset.models.core import Database
-from superset.views import core as views
+from superset.views import core, geocoding
 
 from .base_tests import SupersetTestCase
 
 
 class GeocodingTests(SupersetTestCase):
-    superset = views.Superset()
+    superset = core.Superset()
 
     def __init__(self, *args, **kwargs):
         super(GeocodingTests, self).__init__(*args, **kwargs)
@@ -37,6 +37,37 @@ class GeocodingTests(SupersetTestCase):
 
     def tearDown(self):
         self.logout()
+
+    def test_menu_entry_geocode_exist(self):
+        url = "/dashboard/list/"
+        dashboard_page = self.get_resp(url)
+        assert "Geocode Addresses" in dashboard_page
+
+    def test_geocode_adresses_view_load(self):
+        url = "/geocoder/geocoding"
+        form_get = self.get_resp(url)
+        assert "Geocode Addresses" in form_get
+
+    def test_get_columns(self):
+        url = "/geocoder/geocoding/columns"
+
+        table = db.session.query(SqlaTable).first()
+        table_name = table.table_name
+        columns = reflection.Inspector.from_engine(db.engine).get_columns(table_name)
+
+        data = {"tableName": table_name}
+        response = self.get_resp(url, json_=data)
+        assert columns[0].get("name") in response
+
+    def test_get_invalid_columns(self):
+        url = "/geocoder/geocoding/columns"
+        table_name = "no_table"
+
+        data = {"tableName": table_name}
+        response = self.get_resp(url, json_=data)
+
+        message = "No columns found for table with name {0}".format(table_name)
+        assert message in response
 
     # def test_add_lat_lon_columns(self):
     #     table = db.session.query(SqlaTable).first()
@@ -51,7 +82,7 @@ class GeocodingTests(SupersetTestCase):
     #     columns = reflection.Inspector.from_engine(db.engine).get_columns(table_name)
     #     number_of_columns_before = len(columns)
     #
-    #     views.Superset()._add_lat_lon_columns(
+    #     geocoding.Geocoder()._add_lat_lon_columns(
     #         table_name, lat_column_name, lon_column_name
     #     )
     #
@@ -75,7 +106,7 @@ class GeocodingTests(SupersetTestCase):
         first_column_name = "num"
         second_column_name = "state"
 
-        views.Superset()._insert_geocoded_data(
+        geocoding.Geocoder()._insert_geocoded_data(
             table_name, first_column_name, second_column_name, selected_columns, data
         )
         result = db.engine.execute(
@@ -83,34 +114,3 @@ class GeocodingTests(SupersetTestCase):
         )
         for row in result:
             assert row in data
-
-    def test_menu_entry_geocode_exist(self):
-        url = "/dashboard/list/"
-        dashboard_page = self.get_resp(url)
-        assert "Geocode Addresses" in dashboard_page
-
-    # def test_geocode_adresses_view_load(self):
-    # url = "/superset/geocoding"
-    # form_get = self.get_resp(url)
-    # assert "Geocode Addresses" in form_get
-
-    def test_get_columns(self):
-        url = "/superset/geocoding/columns"
-
-        table = db.session.query(SqlaTable).first()
-        table_name = table.table_name
-        columns = reflection.Inspector.from_engine(db.engine).get_columns(table_name)
-
-        data = {"tableName": table_name}
-        response = self.get_resp(url, json_=data)
-        assert columns[0].get("name") in response
-
-    def test_get_invalid_columns(self):
-        url = "/superset/geocoding/columns"
-        table_name = "no_table"
-
-        data = {"tableName": table_name}
-        response = self.get_resp(url, json_=data)
-
-        message = "No columns found for table with name {0}".format(table_name)
-        assert message in response
